[PhpConfigPrinter] Properly import function calls
I noticed that the `service` function would not be imported in my setup.

It would be imported as:
```php
use Symfony\Component\DependencyInjection\Loader\Configurator\service;
```
instead of
```php
use function Symfony\Component\DependencyInjection\Loader\Configurator\service;
```

I checked the code, and it turns out that it's a bit naive. It has a special rule
for when the function is called `ref` (the old `service` function). In other cases
it imports it as function when the function exists.

Instead of doing this, why not use the nodes to know the context of the import?

The `ImportFullyQualifiedNamesNodeVisitor` keeps track of the context in where it finds a
`FullyQualified` node. When the `FullyQualified` node is used inside a `FuncCall` expression
it will import the FQCN as function.