Ensure only .old at end is picked up
Added if condition to ensure it only matches when .old is at the end of the filepath, avoiding matches of a file name "test.old.csv" (They already shouldn't match theoretically, better to prevent it all together. Additionally it prevents having to do a filesystem check for the file in those scenarios. 

Use variables to satisfy PHPStan

Forcibly cast return of strrpos to int (since it can return false, when this happens it will become 0 and not match in the if condition)