Replace rtrim with substr method (#3607)
* Replace rtrim with substr method

Rtrim removes the characters rather than the word, resulting in edge cases such .xml.old being reduced to .xm 

This fix should be near equal performance wise, without rtrim drawbacks.

* Ensure only .old at end is picked up

Added if condition to ensure it only matches when .old is at the end of the filepath, avoiding matches of a file name "test.old.csv" (They already shouldn't match theoretically, better to prevent it all together. Additionally it prevents having to do a filesystem check for the file in those scenarios. 

Use variables to satisfy PHPStan

Forcibly cast return of strrpos to int (since it can return false, when this happens it will become 0 and not match in the if condition)

* Create xml.old file for test

* Create .xml test for test

* Test should now return 2 files

* Fix typo

Co-authored-by: Tomas Votruba <tomas.vot@gmail.com>