Add reproduction for broken package ignore
The monorepo-builder configuration should ignore the package at
`packages/monorepo-builder/ignore/composer.json` but the ExcludeFile
iterator will only match against a sub-path.

Running `composer merge` demonstrates that the invalid package contained
in the test composer.json ends up in our root composer.json which is
undesirable.

Similarly using a relative path as argument to
`packageDirectoriesExcludes` would not work either because the
validation in that function requires the file to exist and it won't be
able to properly resolve a relative file path.